Update test_guess_the_word.py
Condensed 3 tests into 2.... Used to have a test function for correct guesses and one for incorrect guesses.  Now it tests the random selection of a word, and then another function tests 2 correct guesses and 1 incorrect guess to check the word updates guessed_letters list, and the countdown of number of guesses allowed.